requirements.txt
google-api-python-client
youtube-dl
PyYAML
beautifulsoup4
sqlalchemy
